chore: manage natives locally (#111)
* feature: lending supply assets (#94)

* feat: add abi fetch and format for supply

* Chore:tidy up the fetch data for aave available positions

* feat: fetch correctly fetches ticker name APY colateral and image

* feat: make sure all logos are circles, ensure loading is done on refresh of chain and chain change

* feat: make sure all logos are circles, ensure loading is done on refresh of chain and chain change

* feat: refactor refresh

* feat: refactor small supply modal

* feat: refactor small supply modal

* chore: add unkown to interactions

* chore: add unkown to interacts

* chore: turn off lending page visibility

* chore:refactor to remove global vars and lean on the sdk instead

* chore: clean up comments based on teammates review clean up react move static decs to config

* chore: move supported chain configs into its own aave config

* chore: move configs remove DRY

* chore: clean up moving more code to config

* chore: manage native tokens in local repo

---------

+++ vercel-build.sh
@@ -5,9 +5,9 @@ TEMP_DIR=$(mktemp -d)
 git clone https://x-access-token:${GITHUB_TOKEN_FETCHER_TOKEN}@github.com/altverseweb3/token-fetcher.git "$TEMP_DIR/token-fetcher"
 
 # Remove existing tokens except mono and branded (directories and files)
-echo "Removing existing tokens content except mono and branded..."
+echo "Removing existing tokens content except mono, branded, and natives..."
 for item in public/tokens/*; do
-  if [[ "$item" != "public/tokens/mono" && "$item" != "public/tokens/branded" ]]; then
+  if [[ "$item" != "public/tokens/mono" && "$item" != "public/tokens/branded" && "$item" != "public/tokens/native" ]]; then
     rm -rf "$item"
   fi
 done
